Move function to utils and align header search
